Fixed load factor implementations by using floating point arithmetic

+++ src/main/java/edu/mit/ita/bags/DynamicArrayBag.java
@@ -120,7 +120,7 @@ public class DynamicArrayBag<T> implements Collection<T> {
     }
 
     private float loadFactor() {
-        return size() / capacity();
+        return (float)size() / (float)capacity();
     }
 
     @SuppressWarnings({"unchecked", "ManualArrayCopy"})
